Revamp progress bars, allowing use outside of CLI
Fixes #259
